Рефакторинг CircularCloudLayouterTest.
1. Тесты с CircularCloudLayouterMainRequirementsTest добавлены в CircularCloudLayouterTest.
2. Рефакторинг CircularCloudLayouterTest.
